🚧 started implementing Security filtering. Not yet tested within Spring Boot context

+++ src/main/java/lu/pokevax/business/vaccine/administered/AdministeredVaccineRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdministeredVaccineRepository extends JpaRepository<AdministeredVaccineEntity, Long> {
+public interface AdministeredVaccineRepository extends JpaRepository<AdministeredVaccineEntity, Integer> {
 
 }
